Restore the ability for plugins to set Store headers
The Auth plugin in the v1.2.x series relies on the Store header parsing
the value of a jQuery `annotator:headers` field on the annotation
element in order to set the auth token header on requests to the store.

This functionality got lost in the refactoring of the Store plugin,
which I don't believe is a bad thing (for one, the implementation fell
far short of the "Tell, Don't Ask"[1] design philosophy).

We still need a way to get headers into the Store plugin, at least for
now, so this commit adds a `setHeader` method to the Store plugin, and
updates the Auth plugin to use that.

[1]: http://pragprog.com/articles/tell-dont-ask